Remove the "Shenango" prefix from AsyncUDPSocket and AsyncUDPServerSocket
